Separate PDP and MadCatz configurations
