+++ src/main.cpp
@@ -1,60 +1,13 @@
-#include <MIDI.h>
-#include "RBConfiguration.h"
+// Only uncomment this if you're using MadCatz brain
+// #define MADCATZ_CONFIG
+// Uncomment this if you want to use your drum kit like the toy drum
+// #define USE_RB_DRUM_CONFIG
+#ifndef MADCATZ_CONFIG
+#include "PDPConfig.h"
+#else
+#include "MadCatzConfig.h"
+#endif
 #pragma GCC diagnostic error "-Wextra"
-
-enum output_t
-{
-    FIRST_OUT,
-    OUT_KICK,
-    OUT_PAD_RED,
-    OUT_PAD_YELLOW,
-    OUT_PAD_BLUE,
-    OUT_PAD_GREEN,
-    OUT_CYM_YELLOW,
-    OUT_CYM_BLUE,
-    OUT_CYM_GREEN,
-    LAST_OUT = OUT_CYM_GREEN,
-    NUM_OUT,
-};
-
-struct output_timing_t
-{
-    const unsigned long hold;    // useconds
-    const unsigned long holdOff; // useconds
-};
-
-static constexpr output_timing_t kickTiming = {.hold = KICK_HOLD_ON, .holdOff = KICK_HOLD_OFF};
-static constexpr output_timing_t padTiming = {.hold = PADS_HOLD_ON, .holdOff = PADS_HOLD_OFF};
-static constexpr output_timing_t cymTiming = {.hold = CYM_HOLD_ON, .holdOff = CYM_HOLD_OFF};
-
-struct output_conf_t
-{
-    const int pin;
-    const output_timing_t *timing;
-};
-
-static constexpr output_conf_t outputs[NUM_OUT] = {
-    [FIRST_OUT] = {}, // Not used
-    [OUT_KICK] = {.pin = 2, .timing = &kickTiming},
-    [OUT_PAD_RED] = {.pin = A0, .timing = &padTiming},
-    [OUT_PAD_YELLOW] = {.pin = A1, .timing = &padTiming},
-    [OUT_PAD_BLUE] = {.pin = A2, .timing = &padTiming},
-    [OUT_PAD_GREEN] = {.pin = A3, .timing = &padTiming},
-    [OUT_CYM_YELLOW] = {.pin = 3, .timing = &cymTiming},
-    [OUT_CYM_BLUE] = {.pin = 4, .timing = &cymTiming},
-    [OUT_CYM_GREEN] = {.pin = 5, .timing = &cymTiming},
-};
-
-struct output_state_t
-{
-    output_state_t()
-    {
-        triggeredAt = 0;
-        triggered = false;
-    }
-    unsigned long triggeredAt;
-    bool triggered;
-};
 
 static output_state_t outputStates[NUM_OUT];
 
@@ -81,7 +34,7 @@ static bool outputForNote(const uint8_t &note, output_t &out)
 static void writeOutput(const output_t &out, const bool &state)
 {
     digitalWrite(outputs[out].pin, state);
-    outputStates[out].triggered = (out == OUT_KICK ? !state : state);
+    outputStates[out].triggered = ReturnTriggeredState(out, state);
 }
 
 static void noteOn([[maybe_unused]] const uint8_t &channel, const uint8_t &note, const uint8_t &velocity)
@@ -97,7 +50,7 @@ static void noteOn([[maybe_unused]] const uint8_t &channel, const uint8_t &note,
     auto t = millis();
     if (t - st->triggeredAt > io->timing->hold + io->timing->holdOff)
     {
-        writeOutput(out, (out == OUT_KICK ? LOW : HIGH));
+        writeOutput(out, ActivatePin(out));
         st->triggeredAt = t;
     }
 }
@@ -109,7 +62,8 @@ void setup()
     for (int out = FIRST_OUT; out < NUM_OUT; out++)
     {
         pinMode(outputs[out].pin, OUTPUT);
-        writeOutput(static_cast<output_t>(out), (out == OUT_KICK ? HIGH : LOW));
+        output_t castOut = static_cast<output_t>(out);
+        writeOutput(castOut, DeactivatePin(castOut));
     }
     MIDI.begin(MIDI_CHANNEL);
 }
@@ -130,6 +84,9 @@ void loop()
     for (auto out = 0; out < NUM_OUT; out++)
     {
         if (outputStates[out].triggered && t - outputStates[out].triggeredAt > outputs[out].timing->hold)
-            writeOutput((output_t)out, (out == 1 ? HIGH : LOW));
+        {
+            output_t castOut = static_cast<output_t>(out);
+            writeOutput(castOut, DeactivatePin(castOut));
+        }
     }
 }
